Fix no context issue
Context is fetched for a message by using timestamp ranges using
the hit's timestamp as the midpoint. Even though our hit is
guaranteed to be in that range, sometimes elastic doesn't return
the hit with its context (sad). If not, now the hit manually gets
inserted into the context list.